# Improve xunit perf, in particular for Theory test case discovery