--- a/src/xunit.execution/Sdk/Utility/CommonTasks.cs
+++ b/src/xunit.execution/Sdk/Utility/CommonTasks.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Xunit.Sdk
+{
+    internal static class CommonTasks
+    {
+        internal static readonly Task Completed = Task.FromResult(0);
+    }
+}