… src/xunit.execution/Sdk/Frameworks/XunitTestFrameworkDiscoverer.cs
… src/xunit.execution/Sdk/Reflection/ReflectionAssemblyInfo.cs
… src/xunit.execution/Sdk/Reflection/ReflectionAttributeInfo.cs
… src/xunit.execution/Sdk/Reflection/ReflectionMethodInfo.cs
… src/xunit.execution/Sdk/Reflection/ReflectionTypeInfo.cs
… src/xunit.execution/Sdk/Reflection/Reflector.cs
… src/xunit.execution/Sdk/TypeUtility.cs
… src/xunit.execution/Sdk/Utility/CollectionExtensions.cs