Over at https://github.com/dotnet/corefx, we're starting to use [Theory]ies more aggressively, and we're contemplating some usage that could result in tens of thousands of entries for a given test suite.  Today, this results in some significant performance penalties.

For example, given this test with 40,000 entries:
```C#
public static IEnumerable<object[]> LotsOfData
{
    get
    {
        for (int i = 0; i < 40000; i++)
            yield return new object[] { i, "This is a some data.", "And this is some more data." };
    }
}

[Theory]
[MemberData("LotsOfData")]
public static void TheoryTestWithLotsOfData(int i, string data, string moreData)
{
}
```
on my machine it takes xunit ~30 seconds to discover and run the tests, spending ~21 seconds on discovery and then ~9 seconds on test execution.

I spent some time iteratively profiling for both CPU usage and memory allocations and eliminating the top offenders, mostly low-hanging fruit.  With the commits in this PR, the above example on my machine has discovery time drop from ~21 seconds to ~0.25 seconds, and test execution time drop from ~9 seconds to ~8 seconds, for a total reduction from ~30 seconds to ~8 seconds.

The individual commits have more detail on the exact changes made.

This PR makes a sizeable dent in allocations, but there's still lots of work that can be done to improve things further.  Just as an example, here's a table showing the top-allocated objects, and how many times per test they're being allocated for my earlier example (so multiply each number by 40,000 to get the appx number of allocations for all of the tests together).  I've only shown types of objects allocated 15 or more times per test:

| Type | Number Before | Number After |
| --- | --- | --- |
| String |  278  | 51 |
| Func`2 | 220 | 0  |
| RuntimeMethodInfoStub | 68 | 8  |
| IntPtr | 61 |  1 |
| String[] | 54 |  8 |
| Char[] | 43 | 0  |
| List`1 |  39  | 22 |
| Object[] |  39 | 15 |
| Int[] | 36 | 2  |
| Stack`1 | 23 | 1 |
| CustomAttributeRecord[] | 20  | 14 |
| XunitSerializationTriple | 20 |  0 |
| Attribute[] | 19  |  9 |
| Byte[] | 18 |  0 |
| StringBuilder | 16 | 0  |
| SZGenericArrayEnumerator | 15 |  5 |

There's a very long tail of object types that get allocated twice or once per test as well, but the list was too long to type out here :smile:.