@stephentoub 2015-06-20
ps I should mention that this was all done using the *.dnx assemblies, where AppDomains aren't used.  When looking at perf using the *.desktop assemblies, the biggest remaining cost by far appears to come from the use of .NET Remoting, resulting in huge numbers of allocations as well.