@sharwell 2015-06-20
I've been just starting to look into this due to growing performance concerns for the StyleCopAnalyzers project.

:question: Can you compare the before/after of this on the discovery and execution of the StyleCopAnalyzers project?

:question: Have you tested the performance inside the IDE or the command line? Does this change result in similar performance benefits for both cases?