@bradwilson 2015-06-21
Thanks for the PR. We'll evaluate it when we get together next.