@stephentoub 2015-06-21
> Thanks for the PR. We'll evaluate it when we get together next.

Thanks, Brad.

> Can you compare the before/after of this on the discovery and execution of the StyleCopAnalyzers project?

I ran the tests with the changes, and it didn't really make a measurable difference, maybe a few seconds (out of the two minutes it took to run the tests).  But, as I mentioned in my "p.s." above, with the desktop bits a lot of time is being spent in .NET Remoting-related efforts, and these StyleCopAnalyzers tests are using the desktop bits.  I just profiled the tests briefly, and during the discovery phase at least 80% of the time and 50% of the allocations are spent related to the .NET Remoting stack.