@bradwilson 2015-06-21
@stephentoub @sharwell Theoretically, once this is available, it should give better performance in desktop at the expense of locking the DLLs while the tests run (i.e., appropriate for command-line execution, but not for in-editor execution): https://github.com/xunit/xunit/issues/319