@sharwell 2015-06-21
@stephentoub Thank you. That confirms the results of the testing I performed locally as well. Near the top of my list to experiment with is using a WCF service instead of remoting for the communication layer. For me it's very *very* early stage though.