@bradwilson 2015-06-22
@stephentoub BTW, I feel like there's something else I should mention.

If you're really in a situation where you are returning 40k data rows, then pre-enumeration really doesn't make sense. Aside from being stupidly expensive, it wouldn't be realistic to display them individually to let you run them individually. You can use the `DisableDiscoveryEnumeration` property of `[MemberData]` to turn off pre-enumeration in that scenario.

https://github.com/xunit/xunit/blob/3084c9b60a1207be6c30606441956bb7a99757f6/src/xunit.core/MemberDataAttributeBase.cs#L34