@stephentoub 2015-06-22
Thanks, @bradwilson.  I think the changes in this PR are goodness regardless, but it definitely sounds like we should also be turning off pre-enumeration in general.  Is there a flag that can be specified to do so for all [Theory]s, rather than needing to do it for individual uses of [MemberData]?