@bradwilson 2015-06-22
@stephentoub I agree that the PR is generally useful. We'll definitely be looking at it soon. :)

Today, there is a global flag to turn off pre-enumeration via configuration (see http://xunit.github.io/docs/configuration-files.html); however, you're using DNX, where configuration files aren't yet supported (since the initial version was built on `App.config` files, a concept which DNX does not have).

I've been contemplating how to come up with sensible defaults for this stuff. I have a few notions:

* The most obvious: get a configuration story for DNX. :)
* Pre-enumeration is a UI-centric feature, so runners which just run tests without letting users choose them from a UI should probably never pre-enumerate.
* Pre-enumeration should probably stop once a particular test hits a cap of data items. (25? 50? 100?)

Thoughts on these or other strategies?