@stephentoub 2015-06-22
Thanks, Brad.  Those strategies make sense to me, in particular only pre-enumerating when using a relevant runner (in which case it seems like the default should really be to not pre-enumerate and particular runners can opt to turn it on?)  I don't have a good sense for the cap... seems like you'll probably hit cases where no matter what value you choose, someone will want a higher cap, so you'd probably need it to be configurable.