@bradwilson 2015-06-28
Thanks.

I squashed and rebased in 82e719db5dfa7e31ce31b4b169065e147dea118d. Also, I turned off theory pre-enumeration for the Console and MSBuild runners in 07178782a6ea455ae42f1fa6f4b3bfec935cd587, which should help a little bit more. :)