@stephentoub 2015-06-28
Terrific! Thanks, Brad.